service/ingestion/requirements.txt
grpcio
protobuf
minio
pymongo